+++ jib-gradle-plugin/CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 
 - Throw `IllegalArgumentException` with error message instead of `NullPointerException` when `jib.to.tags` is set to collection containing a `null` value. ([#2760](https://github.com/GoogleContainerTools/jib/issues/2760))
 - Fixed `NullPointerException` during input validation (in Java 9+) when configuring Jib parameters using certain immutable collections (such as `List.of()`). ([#2702](https://github.com/GoogleContainerTools/jib/issues/2702))
+- Fixed an issue that configuring `jib.from.platforms` was always additive to the default `amd64/linux` platform. ([#2783](https://github.com/GoogleContainerTools/jib/issues/2783))
 
 ## 2.5.0